back-end/requirements.txt
flask>=2.0.0
rdflib>=6.0.0
flask-cors>=3.0.10
python-dotenv>=0.19.0